% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{avgVecList}
\alias{avgVecList}
\title{Vector Average function}
\usage{
avgVecList(x = list(c(1, 2, 3), c(4, 5, 6)))
}
\arguments{
\item{x}{\code{list}}
}
\description{
Vector Average function
}
\details{
More informations.
}
\author{
Jacek
}
