% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.r
\name{hello}
\alias{hello}
\title{Hello}
\usage{
hello(name = "world")
}
\arguments{
\item{name}{It defines to whom it says hello.}
}
\value{
It returns the text of saying hello.
}
\description{
Says hello
}
\examples{
hello("me") can be used to say hello to me.

}
