% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.r
\name{result}
\alias{result}
\title{Result}
\usage{
result(x1, x2, operator)
}
\arguments{
\item{x1}{The first variable of the formula.}

\item{x2}{The second variable of the formula.}

\item{operator}{The operator.}
}
\value{
It returns the result of the formula.
}
\description{
Calculates the result
}
\examples{
result(5, 10, "+")

}
