% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shp_to_gpkg.R
\name{shp_to_gpkg}
\alias{shp_to_gpkg}
\title{read the shp file
add the shapefile to a gpkg file
transform the shp file into the crs=4326}
\usage{
shp_to_gpkg(shapefile, gpkgfile)
}
\description{
read the shp file
add the shapefile to a gpkg file
transform the shp file into the crs=4326
}
