% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Write here a title of your function}
\usage{
hello(name = "world")
}
\arguments{
\item{name}{The name that should be greeted.}
}
\value{
Output stdout.
}
\description{
Write here a short description of your function.
}
\examples{
in Python: print("Hello World")

}
